Evaluation is int again & less ordering
Moves only need ordering in search, not always when generated

+++ Piece.h
@@ -23,6 +23,6 @@ public:
 	static short getOppositeColor(short color);
 	static std::string name(short piece);
 	static char toChar(short piece);
-	static float getPieceValue(short piece);
+	static int getPieceValue(short piece);
 };
